gnr: R companion to the Global Nutrition Report

The {gnr} package is a companion package to the Global Nutrition Report primarily for accessing the data used in the report and to reproduce the data visualisation produced in the report. This package contains functions that allow users to interrogate the same data and produce customised charts and other visualisation products.

See also

Author

Maintainer: Ernest Guevarra ernest@guevarra.io (ORCID) [copyright holder]