Description of variables that occurs in the Global Nutrition Report (GNR) dataset

Description of variables that occurs in the Global Nutrition Report (GNR) dataset

Usage

indicator_description

Format

A data.frame with 709 rows and 3 columns:

section

Section of the report

variable

Data variable

description

Description of variable

Source

Global Nutrition Report

Examples

indicator_description
#> # A tibble: 709 × 3
#>    section                variable          description                         
#>    <chr>                  <chr>             <chr>                               
#>  1 Adult nutrition status adult_anemia_1999 Anaemia among women of reproductive…
#>  2 Adult nutrition status adult_anemia_2000 Anaemia among women of reproductive…
#>  3 Adult nutrition status adult_anemia_2001 Anaemia among women of reproductive…
#>  4 Adult nutrition status adult_anemia_2002 Anaemia among women of reproductive…
#>  5 Adult nutrition status adult_anemia_2003 Anaemia among women of reproductive…
#>  6 Adult nutrition status adult_anemia_2004 Anaemia among women of reproductive…
#>  7 Adult nutrition status adult_anemia_2005 Anaemia among women of reproductive…
#>  8 Adult nutrition status adult_anemia_2006 Anaemia among women of reproductive…
#>  9 Adult nutrition status adult_anemia_2007 Anaemia among women of reproductive…
#> 10 Adult nutrition status adult_anemia_2008 Anaemia among women of reproductive…
#> # ℹ 699 more rows